refactor: 함수 분리 - player.js의 경우 player 함수 안에서 호출하는 validation, compare은 같은 파일에 작성
